fastapi-practice/requirements.txt
fastapi
uvicorn

sqlalchemy
passlib
bcrypt
python-multipart
python-jose
python-dotenv